requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.20
django-classy-tags==4.1.0
django-cms==4.1.5
django-formtools==2.5.1
django-localflavor==4.0
django-sekizai==4.1.0
django-storages==1.14.6
django-theherk-events==1.8.1
django-theherk-resources==1.6.1
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
kombu==5.5.2
lxml==5.3.1
minio==7.2.15
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
python-dateutil==2.9.0.post0
python-stdnum==1.20
redis==5.2.1
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions>=4
tzdata==2025.2
urllib3<1.27
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
boto3==1.20.0
gunicorn
python-dotenv
pytz